Correct name of src folder
The Arduino library framework allows the library code files to be stored under a subfolder of the library. This library
must have the exact name `src`.

Previously, the folder was incorrectly named `scr`, which caused the code files stored under the folder to not be
recognized by Arduino development software.